refactor: Restructure bot with modular handler architecture and comprehensive test suite
Refactored the monolithic bot.py (2,103 lines) into a clean orchestrator pattern with domain-separated handlers, reducing complexity and improving maintainability.

Key Changes:
- Reduced bot.py from 2,103 lines to 152 lines (93% reduction)
- Split 590-line button_callback() into 5 domain-specific callback handlers
- Created modular handler architecture with clear separation of concerns
- Added 56 comprehensive tests covering critical functionality (~75% coverage)

New Structure:
- handlers/user_commands.py - User-facing commands
- handlers/mood_commands.py - Mood tracking and stats
- handlers/goal_commands.py - Goal management
- handlers/admin_commands.py - Admin operations
- handlers/message_handler.py - Text message processing
- handlers/callbacks/ - Callback routing and domain handlers
- tests/ - Complete test suite with fixtures and mocks

Benefits:
- Improved maintainability (single-responsibility modules)
- Enhanced testability (isolated handler testing)
- Better scalability (parallel development possible)
- Reduced cognitive load (200-600 lines per file vs 2,103)

Testing:
- 56 tests across 6 test files
- pytest infrastructure with async support
- Comprehensive fixtures for mocking
- Integration and unit test coverage

Documentation:
- Updated CLAUDE.md with new architecture overview
- Created TEST_SUMMARY.md with testing guide
- Updated TODO.md to mark content strategy complete
