+++ TODO.md
@@ -60,29 +60,16 @@ This file tracks cleanup and improvement tasks for the Motivator Bot project.
 
 ---
 
-### 🟡 Content Management Strategy
+### ✅ Content Management Strategy
 **Issue**: Unclear content management approach - both hardcoded and database import system exist
 
-**Current State**:
-- Content is hardcoded in `src/content.py` (177 lines)
-- `scripts/import_content.py` exists (213 lines) but not integrated
-- `config/content_example.json` example file exists
-- Database has content tables but may not be used
+**Resolution**: Database-driven approach implemented
+- [x] Migrate all content from `content.py` to database
+- [x] Integrate `import_content.py` into main workflow
+- [x] Create admin interface for content management
+- [x] Remove hardcoded content arrays from `content.py`
 
-**Decision Required**: Choose ONE approach
-
-**Option A - Keep Hardcoded** (if content is stable):
-- [ ] Remove or document `import_content.py` as legacy tool
-- [ ] Update `content_example.json` to match actual content in code
-- [ ] Add inline documentation to `content.py`
-
-**Option B - Use Database Import** (if content changes frequently):
-- [ ] Migrate all content from `content.py` to database
-- [ ] Integrate `import_content.py` into main workflow
-- [ ] Create admin interface for content management
-- [ ] Remove hardcoded content arrays from `content.py`
-
-**Estimated Effort**: 2-4 hours
+**Completed**: 2025-11-02
 **Impact**: Medium (content flexibility, ease of updates)
 
 ---
