+++ pytest.ini
@@ -1,0 +1,28 @@
+[pytest]
+# Pytest configuration for Motivator Bot
+
+# Test discovery
+python_files = test_*.py
+python_classes = Test*
+python_functions = test_*
+
+# Output options
+addopts =
+    -v
+    --tb=short
+    --strict-markers
+    --disable-warnings
+    -p no:warnings
+
+# Test paths
+testpaths = tests
+
+# Markers
+markers =
+    unit: Unit tests (fast, isolated)
+    integration: Integration tests (slower, uses real components)
+    slow: Slow running tests
+    asyncio: Async tests
+
+# Asyncio configuration
+asyncio_mode = auto
