+++ src/handlers/__init__.py
@@ -1,0 +1,9 @@
+"""
+Handlers package for Motivator Bot.
+
+Contains command handlers and callback handlers split by domain.
+"""
+
+from .base import BaseHandler
+
+__all__ = ['BaseHandler']
